Refactor UniProt client into dedicated module
